// components/nftmarketplace/interface.ts
export type NFTCardType = {
    ownername: string;
    title: string;
    src: string;
    owneravatar: string;
    price: number;
    highestbid: number;
    date: string;
};

export type CollectionType = {
    src: string;
    small: any;
    ownername: string;
    name: string;
    owneravatar: string;
};